Add files via upload
This update is a proof of concept test to show items being added to cart.  Only works with the first product on the stickers page and only displays the item, price, and quantity in a text label on the bottom of the shopping cart page.  This is strictly using sessions(not using database at all).